FS-5062 - Moving tests against build_form_json out of test_integration and into test_generate_form unit test module
